fix: ws in production

+++ prod.sh
@@ -27,8 +27,8 @@ echo "${GREEN}********************"
 echo "${GREEN}* Build successful *"
 echo "${GREEN}********************"
 
-[ -d "${CLIENT}"] && rm -R $CLIENT
-[ -d "${SERVER}" ] && rm -R $SERVER
+[ -d "$CLIENT" ] && rm -R $CLIENT
+[ -d "$SERVER" ] && rm -R $SERVER
 
 mkdir $CLIENT
 mkdir $SERVER
@@ -37,7 +37,7 @@ echo "'${GREEN}'copying client and server bundles"
 cp -R dist/client/* $CLIENT
 cp -R dist/server/* $SERVER
 
-rm /tmp/out
+[ -f "/tmp/out" ] && rm /tmp/out
 echo "#!/usr/bin/env node"|cat - $SERVER/main.js > /tmp/out && mv /tmp/out $SERVER/main.js
 
 chmod +x $SERVER/main.js
